pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Vita <vitay2715@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.11"
requests = "2.31.0"
djangorestframework = "3.15.1"
psycopg2-binary = "2.9.9"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
django-cors-headers = "4.3.1"
drf-spectacular = "^0.27.2"
pillow = "^10.3.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}

[tool.poetry.group.dev.dependencies]
flake8 = "7.0.0"
isort = "5.13.2"
black = "24.3.0"


[[source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
